# agent-manager-observer/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agent Manager Observer API
  description: |
    Observer Service-backed API for listing, exporting, and fetching distributed traces,
    component/build logs, and resource metrics.

    Endpoints use human-readable `organization`, `project`, `agent`, and `environment` identifiers.
    Lazy loading is supported: the span list endpoint returns summaries only; individual span
    attributes are fetched on demand via the span detail endpoint.

    All endpoints require the same Bearer JWT as the Agent Manager Service console
    (validated via JWKS / local-dev mode).
  version: 1.0.0
  contact:
    name: WSO2 LLC
    url: https://www.wso2.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:9098
    description: Local development
  - url: https://observer.example.com
    description: Production (replace with your deployment URL)
...
paths:
...
  /api/v1/traces:
    get:
      tags:
        - traces
      summary: List traces
      description: |
        Returns paginated trace overviews for a component. Each trace includes input, output, token usage,
        and status extracted from the root span — the same richness as the v1 list endpoint.
        Uses the Observer Service instead of OpenSearch. No offset; use `limit` for page size.

        The filter parameters (`status`, `minDurationMs`, `minTokens`, `minSpanCount`, `model`,
        `conversationId`) are optional and combine with AND. They are applied to the fetched page
        after enrichment, so a filtered response can hold fewer than `limit` traces. With no filter
        set the request costs the same upstream calls as an unfiltered list.
      operationId: listTraces
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/organization"
        - $ref: "#/components/parameters/project"
        - $ref: "#/components/parameters/agent"
        - $ref: "#/components/parameters/environment"
        - $ref: "#/components/parameters/startTimeRequired"
        - $ref: "#/components/parameters/endTimeRequired"
        - name: limit
          in: query
          required: false
          description: Page size (default 10, max 1000)
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 10
        - $ref: "#/components/parameters/sortOrder"
        - name: include
          in: query
          required: false
          style: form
          explode: false
          description: |
            Opt-in span-derived fields to fill on every trace, as a comma-separated
            list. The parameter may be repeated. Unknown values return 400.

            - `models`: read model names from the trace's span list, requested with
              inline attributes. It adds one span-list call for traces that otherwise
              need none, and makes the span-list call heavier for the rest. Traces
              above the span-count threshold return no `models` either way.
          schema:
            type: array
            items:
              type: string
              enum: [models]
        - $ref: "#/components/parameters/traceStatusFilter"
        - $ref: "#/components/parameters/minDurationMsFilter"
        - $ref: "#/components/parameters/minTokensFilter"
        - $ref: "#/components/parameters/minSpanCountFilter"
        - $ref: "#/components/parameters/modelFilter"
        - $ref: "#/components/parameters/conversationIdFilter"
      responses:
        "200":
          description: Trace overview list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceOverviewListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ObserverUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/traces/export:
    get:
      tags:
        - traces
      summary: Export traces with full span details
      description: |
        Returns complete trace objects for a component including ALL spans, each fully enriched with
        AmpAttributes (LLM data, tool calls, token usage, etc.). Response includes Content-Disposition
        for download. Export limit is capped at 1000 traces.

        `organization`, `project`, `agent`, and `environment` are all required to scope the export
        to a specific agent.
      operationId: exportTraces
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/organization"
        - $ref: "#/components/parameters/project"
        - $ref: "#/components/parameters/agent"
        - $ref: "#/components/parameters/environment"
        - $ref: "#/components/parameters/startTimeRequired"
        - $ref: "#/components/parameters/endTimeRequired"
        - name: limit
          in: query
          required: false
          description: Max traces to export (default 100, max 1000)
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - $ref: "#/components/parameters/sortOrder"
      responses:
        "200":
          description: Export payload
          headers:
            Content-Disposition:
              description: Suggested download filename
              schema:
                type: string
                example: 'attachment; filename="traces-export-20260406-120000.json"'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceExportResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ObserverUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/traces/{traceId}/spans/{spanId}:
    get:
      tags:
        - traces
      summary: Get full span detail
      description: |
        Returns the full span including all OTEL attributes and enriched AmpAttributes
        (semantic kind, input/output, token usage, LLM data, tool calls, etc.).
      operationId: getSpanDetail
      security:
        - bearerAuth: []
      parameters:
        - name: traceId
          in: path
          required: true
          description: Trace identifier
          schema:
            type: string
            example: "3cae024cf613a5f37843e9c6eefa3020"
        - name: spanId
          in: path
          required: true
          description: Span identifier
          schema:
            type: string
            example: "a1b2c3d4e5f60001"
      responses:
        "200":
          description: Full enriched span
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrichedSpan"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ObserverUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Same access token as used for Agent Manager Service API calls.
  parameters:
    organization:
      name: organization
      in: query
      required: true
      description: Organization name
      schema:
        type: string
        example: "default"
    project:
      name: project
      in: query
      required: true
      description: Project name
      schema:
        type: string
    agent:
      name: agent
      in: query
      required: true
      description: Agent name
      schema:
        type: string
    environment:
      name: environment
      in: query
      required: true
      description: Environment name
      schema:
        type: string
...
    startTimeRequired:
      name: startTime
      in: query
      required: true
      schema:
        type: string
        format: date-time
        example: "2026-04-01T00:00:00Z"
    endTimeRequired:
      name: endTime
      in: query
      required: true
      schema:
        type: string
        format: date-time
        example: "2026-04-06T23:59:59Z"
    sortOrder:
      name: sortOrder
      in: query
      required: false
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    traceStatusFilter:
      name: status
      in: query
      required: false
      description: |
        `error` keeps traces whose status reports at least one error; `ok` keeps the rest,
        including traces with no status.
      schema:
        type: string
        enum: [error, ok]
    minDurationMsFilter:
      name: minDurationMs
      in: query
      required: false
      description: Keep traces whose duration is at least this many milliseconds.
      schema:
        type: integer
        format: int64
        minimum: 0
    minTokensFilter:
      name: minTokens
      in: query
      required: false
      description: |
        Keep traces whose `tokenUsage.totalTokens` is at least this value. Traces without
        `tokenUsage` never match, even with `minTokens=0`.
      schema:
        type: integer
        format: int64
        minimum: 0
    minSpanCountFilter:
      name: minSpanCount
      in: query
      required: false
      description: Keep traces with at least this many spans.
      schema:
        type: integer
        format: int64
        minimum: 0
    modelFilter:
      name: model
      in: query
      required: false
      description: |
        Keep traces where any entry of `models` equals this value exactly. Implies
        `include=models`, so the server fills `models` on every trace and the request
        carries that cost. Traces above the span-count threshold have no `models` and
        never match.
      schema:
        type: string
        example: gpt-4o
    conversationIdFilter:
      name: conversationId
      in: query
      required: false
      description: Keep traces whose `conversationId` equals this value exactly.
      schema:
        type: string
        example: conv-42
  responses:
    BadRequest:
      description: Bad request — missing or invalid query parameter
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ObserverUnavailable:
      description: Observer Service not configured (OPENCHOREO_OBSERVER_URL unset)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Short error code or literal (e.g. error, Unauthorized)
        message:
          type: string
          description: Human-readable detail
...
    TraceExportResponse:
      type: object
      required:
        - traces
        - totalCount
        - truncated
      properties:
        traces:
          type: array
          items:
            $ref: "#/components/schemas/FullTrace"
        totalCount:
          type: integer
        truncated:
          type: boolean
          description: True if any trace hit the MaxSpansPerRequest cap (10 000 spans)

    # ── schemas ─────────────────────────────────────────────────────────────
    TraceOverview:
      type: object
      description: |
        Single trace overview as returned by GET /api/v1/traces.
        Includes root-span-derived input, output, and token usage.
      required:
        - traceId
        - rootSpanId
        - rootSpanName
        - rootSpanKind
        - startTime
        - endTime
        - durationInNanos
        - spanCount
      properties:
        traceId:
          type: string
        rootSpanId:
          type: string
        rootSpanName:
          type: string
        rootSpanKind:
          type: string
          description: Semantic span type (llm, tool, agent, chain, etc.)
          example: llm
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        durationInNanos:
          type: integer
          format: int64
        spanCount:
          type: integer
        tokenUsage:
          $ref: "#/components/schemas/TokenUsage"
        status:
          $ref: "#/components/schemas/TraceStatus"
        input:
          nullable: true
          oneOf:
            - type: string
            - type: object
            - type: array
              items: {}
        output:
          nullable: true
          oneOf:
            - type: string
            - type: object
            - type: array
              items: {}
        models:
          type: array
          description: |
            Distinct model names across the trace's LLM spans, first-seen order.
            Only present when `include=models`, or when the LLM spans were
            fetched for other reasons (to fill `input`, `output` or `tokenUsage`).
            Absent otherwise, and absent when no LLM span names a model.
          items:
            type: string
          example: ["gpt-4o", "gpt-4o-mini"]
        conversationId:
          type: string
          description: gen_ai.conversation.id of the root span.
          example: conv-42
    TraceOverviewListResponse:
      type: object
      required:
        - traces
        - totalCount
        - lookedBackTo
        - truncated
      properties:
        traces:
          type: array
          items:
            $ref: "#/components/schemas/TraceOverview"
        totalCount:
          type: integer
          description: |
            Upstream trace count for the window when no filter is set; the number of
            matching traces in `traces` when any filter is set.
        lookedBackTo:
          type: string
          format: date-time
          description: |
            How far back in the window the server examined traces, in UTC with sub-second
            precision so it can be passed back as a time bound without gaps. Currently always
            the requested `startTime`.
          example: "2026-04-01T00:00:00.5Z"
        truncated:
          type: boolean
          description: |
            True when the server stopped at its examine cap before the window ran out,
            so older matches may exist before `lookedBackTo`. Currently always false.
...
    EnrichedSpan:
      description: |
        Full span as returned by GET /api/v1/traces/{traceId}/spans/{spanId}.
        Includes raw OTEL attributes and enriched AmpAttributes.
      allOf:
        - $ref: "#/components/schemas/Span"
        - type: object
          properties:
            ampAttributes:
              $ref: "#/components/schemas/AmpAttributes"
    TokenUsage:
      type: object
      properties:
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        totalTokens:
          type: integer
    TraceStatus:
      type: object
      properties:
        errorCount:
          type: integer
...
security:
  - bearerAuth: []